pyproject.toml
[tool.poetry]
name = "bedrock-agents"
version = "0.1.0"
description = ""
authors = ["Brian Peterson <brian.peterson@cloudshift.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.148.0"
constructs = ">=10.0.0,<11.0.0"
pinecone-client = "^4.1.2"
python-dotenv = "^1.0.1"
aws-lambda-typing = "^2.20.0"
aws-lambda-powertools = "^2.40.1"
pydantic = "^2.8.2"
requests = "^2.32.3"
streamlit = "^1.36.0"
gradio = "^4.38.1"
pandas = "^2.2.2"
plotly = "^5.23.0"


[tool.poetry.group.dev.dependencies]
boto3 = "^1.34.141"
jupyter = "^1.0.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
black = "^24.4.2"
autoflake = "^2.3.1"
flake8 = "^7.1.0"
watchdog = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
